refactor(tests): remove unused Cypress exports and flatten Playwright imports
Remove dead Cypress helpers (interceptExploreJson, interceptFormDataKey,
clearAllInputs, getChartAliasBySpec, getChartAliasesBySpec, toSlicelike,
ALERT_LIST, REPORT_LIST, LOGIN, REGISTER) and unexport internal-only
functions (interceptExploreGet, getChartGridComponent). Flatten Playwright
test imports to use barrel index files for components/modals,
components/core, and helpers/fixtures. Remove unused clickCardDeleteAction
from ChartListPage.
