+++ superset-frontend/playwright/components/modals/index.ts
@@ -18,6 +18,9 @@
  */
 
 // Specific modal implementations
+export { ChartPropertiesModal } from './ChartPropertiesModal';
+export { ConfirmDialog } from './ConfirmDialog';
 export { DeleteConfirmationModal } from './DeleteConfirmationModal';
 export { DuplicateDatasetModal } from './DuplicateDatasetModal';
+export { EditDatasetModal } from './EditDatasetModal';
 export { ImportDatasetModal } from './ImportDatasetModal';
