+++ superset-frontend/playwright/helpers/fixtures/index.ts
@@ -18,4 +18,9 @@
  */
 
 // Base fixture with test asset cleanup
-export { test as testWithAssets, expect, type TestAssets } from './testAssets';
+export {
+  test,
+  test as testWithAssets,
+  expect,
+  type TestAssets,
+} from './testAssets';
