+++ superset-frontend/playwright/tests/chart/chart-test-helpers.ts
@@ -18,7 +18,7 @@
  */
 
 import type { Page, TestInfo } from '@playwright/test';
-import type { TestAssets } from '../../helpers/fixtures/testAssets';
+import type { TestAssets } from '../../helpers/fixtures';
 import { apiPostChart } from '../../helpers/api/chart';
 import { getDatasetByName } from '../../helpers/api/dataset';
