+++ superset-frontend/playwright/tests/dashboard/dashboard-test-helpers.ts
@@ -18,7 +18,7 @@
  */
 
 import type { Page, TestInfo } from '@playwright/test';
-import type { TestAssets } from '../../helpers/fixtures/testAssets';
+import type { TestAssets } from '../../helpers/fixtures';
 import { apiPostDashboard } from '../../helpers/api/dashboard';
 
 interface TestDashboardResult {
